+++ src/main/java/nextstep/reservation/dao/ReservationDAO.java
@@ -1,0 +1,25 @@
+package nextstep.reservation.dao;
+
+import nextstep.reservation.dto.ReservationDTO;
+
+import java.sql.Date;
+import java.sql.Time;
+
+public interface ReservationDAO {
+
+    String SELECT_BY_DATE_AND_TIME_SQL = "SELECT count(*) FROM reservation WHERE date = ? AND time = ? LIMIT 1";
+
+    String INSERT_SQL = "INSERT INTO reservation(date,time,name,theme_name,theme_desc,theme_price) VALUES(?,?,?,?,?,?)";
+
+    String SELECT_BY_ID_SQL = "SELECT * FROM reservation WHERE id = ?";
+
+    String DELETE_BY_ID_SQL = "DELETE FROM reservation WHERE id = ?";
+
+    Boolean existsByDateAndTime(Date date, Time time) throws RuntimeException;
+
+    Long insert(ReservationDTO dto);
+
+    ReservationDTO getById(Long id);
+
+    Boolean deleteById(Long id);
+}
