+++ src/main/java/nextstep/reservation/dto/ReservationRequest.java
@@ -1,4 +1,4 @@
-package nextstep.dto;
+package nextstep.reservation.dto;
 
 import lombok.Data;
