+++ src/main/java/nextstep/reservation/repository/ReservationRepositoryImpl.java
@@ -1,10 +1,10 @@
-package nextstep.repository;
+package nextstep.reservation.repository;
 
 import lombok.AllArgsConstructor;
-import nextstep.Reservation;
-import nextstep.dao.ReservationDAO;
-import nextstep.dto.ReservationDTO;
-import nextstep.mapper.ReservationMapper;
+import nextstep.reservation.dao.ReservationDAO;
+import nextstep.reservation.domain.Reservation;
+import nextstep.reservation.dto.ReservationDTO;
+import nextstep.reservation.mapper.ReservationMapper;
 import org.springframework.stereotype.Repository;
 
 import java.sql.Date;
@@ -12,15 +12,15 @@ import java.sql.Time;
 
 @AllArgsConstructor
 @Repository
-public class ReservationRepositoryImpl implements ReservationRepository{
+public class ReservationRepositoryImpl implements ReservationRepository {
 
     private final ReservationDAO dao;
 
     @Override
-    public Long insertIfNotExistsDateTime(Reservation reservation){
+    public Long insertIfNotExistsDateTime(Reservation reservation) {
         Date date = Date.valueOf(reservation.getDate());
         Time time = Time.valueOf(reservation.getTime());
-        if(dao.existsByDateAndTime(date, time)) {
+        if (dao.existsByDateAndTime(date, time)) {
             throw new RuntimeException("해당 날짜에는 이미 예약이 존재합니다.");
         }
         ReservationDTO dto = ReservationMapper.INSTANCE.toDto(reservation);
@@ -29,7 +29,7 @@ public class ReservationRepositoryImpl implements ReservationRepository{
     }
 
     @Override
-    public Reservation getById(Long id){
+    public Reservation getById(Long id) {
         ReservationDTO dto = dao.getById(id);
 
         return ReservationMapper.INSTANCE.fromDto(dto);
